Recreation of an insurance policy management service.
The Insurance Company must deploy a Smart Contract
InsuranceFactory (IF) that allows all the management of this practice. Through this
Smart Contract policyholders will be able to buy tokens that comply with the token standard
ERC 20 to be able to pay for services. Customers will have to buy the
tokens with cryptocurrencies (ethers). Naturally, the company must generate and have
in its balance a number of tokens sufficient to be able to sell them to the insured.
Also, through the company's Smart Contract IF, it must be possible to register the
different services with the corresponding price expressed in tokens.
In order to be an insured, each client of the company has to perform a function
to create your Smart Contract.